refactor: uint256 === uint

+++ contracts/WavePortal.sol
@@ -5,7 +5,7 @@ pragma solidity ^0.8.0;
 import "hardhat/console.sol";
 
 contract WavePortal {
-    uint256 totalWaves;
+    uint totalWaves;
 
     constructor() {
         console.log("This is my awesome first contract!");
@@ -16,7 +16,7 @@ contract WavePortal {
         console.log("%s has waved!", msg.sender);
     }
 
-    function getTotalWaves() public view returns (uint256) {
+    function getTotalWaves() public view returns (uint) {
         console.log("We have %d total waves!", totalWaves);
         return totalWaves;
     }
